Temperature and pwm functional
