+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "sdkconfig.h": "c"
+        "sdkconfig.h": "c",
+        "definitions.h": "c"
     }
 }
